requirements.txt
pika==1.3.2
python-dotenv==1.0.0
colorama==0.4.6
flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
eventlet==0.33.3